Separate subtitle download hook
