+++ app/src/main/java/me/iacn/biliroaming/XposedInit.kt
@@ -113,6 +113,7 @@ class XposedInit : IXposedHookLoadPackage, IXposedHookZygoteInit {
                     startHook(TryWatchVipQualityHook(lpparam.classLoader))
                     startHook(AllowMiniPlayHook(lpparam.classLoader))
                     startHook(VideoSubtitleHook(lpparam.classLoader))
+                    startHook(SubtitleDownloadHook(lpparam.classLoader))
                 }
                 lpparam.processName.endsWith(":web") -> {
                     BiliBiliPackage(lpparam.classLoader, param.args[0] as Context)
