+++ app/src/main/java/me/iacn/biliroaming/hook/BangumiPlayUrlHook.kt
@@ -152,7 +152,7 @@ class BangumiPlayUrlHook(classLoader: ClassLoader) : BaseHook(classLoader) {
                         param.result = fixDownloadProto(response)
                     }
                 }
-                VideoSubtitleHook.onEpPlay()
+                SubtitleDownloadHook.onEpPlay()
             }
         }
         "com.bapis.bilibili.pgc.gateway.player.v2.PlayURLMoss".findClassOrNull(mClassLoader)?.run {
@@ -215,7 +215,7 @@ class BangumiPlayUrlHook(classLoader: ClassLoader) : BaseHook(classLoader) {
                         param.result = fixDownloadProto(response)
                     }
                 }
-                VideoSubtitleHook.onEpPlay()
+                SubtitleDownloadHook.onEpPlay()
             }
         }
